Revert "Revert "Merge pull request #28121 from deads2k/invariant-15-extract-run-test""
This reverts commit 759181c63b705b3fa815a627a66f6db82f0d39f1.
