+++ pkg/cmd/openshift-tests/run-upgrade/flags.go
@@ -1,14 +1,13 @@
-package main
+package run_upgrade
 
 import (
 	"fmt"
-
-	"k8s.io/cli-runtime/pkg/genericclioptions"
 
 	"github.com/openshift/origin/pkg/clioptions/clusterdiscovery"
 	"github.com/openshift/origin/pkg/clioptions/iooptions"
 	testginkgo "github.com/openshift/origin/pkg/test/ginkgo"
 	"github.com/spf13/pflag"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
 )
 
 // TODO collapse this with cmd_runsuite
@@ -64,7 +63,6 @@ func (f *RunUpgradeSuiteFlags) ToOptions(args []string) (*RunUpgradeSuiteOptions
 
 	// shallow copy to mutate
 	ginkgoOptions := f.GinkgoRunSuiteOptions
-	ginkgoOptions.SyntheticEventTests = pulledInvalidImages(f.FromRepository)
 	// Upgrade test output is important for debugging because it shows linear progress
 	// and when the CVO hangs.
 	ginkgoOptions.IncludeSuccessOutput = true
