+++ pkg/cmd/openshift-tests/run-upgrade/options.go
@@ -1,4 +1,4 @@
-package main
+package run_upgrade
 
 import (
 	"context"
@@ -7,32 +7,43 @@ import (
 	"path/filepath"
 
 	"github.com/openshift/origin/pkg/clioptions/clusterdiscovery"
+	"github.com/openshift/origin/pkg/clioptions/imagesetup"
 	"github.com/openshift/origin/pkg/clioptions/iooptions"
+	"github.com/openshift/origin/pkg/clioptions/upgradeoptions"
+	"github.com/openshift/origin/pkg/test/ginkgo"
 	testginkgo "github.com/openshift/origin/pkg/test/ginkgo"
 	"github.com/openshift/origin/pkg/version"
+	"github.com/openshift/origin/test/e2e/upgrade"
+	exutil "github.com/openshift/origin/test/extended/util"
 	"github.com/openshift/origin/test/extended/util/image"
+	"github.com/pkg/errors"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"k8s.io/klog/v2"
 	k8simage "k8s.io/kubernetes/test/utils/image"
 )
 
 // TODO collapse this with cmd_runsuite
-type RunSuiteOptions struct {
+type RunUpgradeSuiteOptions struct {
 	GinkgoRunSuiteOptions *testginkgo.GinkgoRunSuiteOptions
 	Suite                 *testginkgo.TestSuite
 
-	FromRepository    string
-	CloudProviderJSON string
+	ToImage        string
+	FromRepository string
+	// I don't see where this is initialized in this flow
+	//CloudProviderJSON string
+
+	TestOptions []string
 
 	CloseFn iooptions.CloseFunc
+
 	genericclioptions.IOStreams
 }
 
-func (o *RunSuiteOptions) TestCommandEnvironment() []string {
+func (o *RunUpgradeSuiteOptions) TestCommandEnvironment() []string {
 	var args []string
 	args = append(args, "KUBE_TEST_REPO_LIST=") // explicitly prevent selective override
 	args = append(args, fmt.Sprintf("KUBE_TEST_REPO=%s", o.FromRepository))
-	args = append(args, fmt.Sprintf("TEST_PROVIDER=%s", o.CloudProviderJSON))
+	//args = append(args, fmt.Sprintf("TEST_PROVIDER=%s", o.CloudProviderJSON))  I don't think we actually have this.
 	args = append(args, fmt.Sprintf("TEST_JUNIT_DIR=%s", o.GinkgoRunSuiteOptions.JUnitDir))
 	for i := 10; i > 0; i-- {
 		if klog.V(klog.Level(i)).Enabled() {
@@ -40,12 +51,41 @@ func (o *RunSuiteOptions) TestCommandEnvironment() []string {
 			break
 		}
 	}
-	args = append(args, "TEST_UPGRADE_OPTIONS=")
+
+	upgradeOptions := upgradeoptions.UpgradeOptions{
+		Suite:       o.Suite.Name,
+		ToImage:     o.ToImage,
+		TestOptions: o.TestOptions,
+	}
+	args = append(args, fmt.Sprintf("TEST_UPGRADE_OPTIONS=%s", upgradeOptions.ToEnv()))
 
 	return args
 }
 
-func (o *RunSuiteOptions) Run(ctx context.Context) error {
+// UpgradeTestPreSuite validates the test options and gathers data useful prior to launching the upgrade and it's
+// related tests.
+func (o *RunUpgradeSuiteOptions) UpgradeTestPreSuite() error {
+	if !o.GinkgoRunSuiteOptions.DryRun {
+		testOpt := ginkgo.NewTestOptions(o.IOStreams)
+		config, err := clusterdiscovery.DecodeProvider(os.Getenv("TEST_PROVIDER"), testOpt.DryRun, false, nil)
+		if err != nil {
+			return err
+		}
+		if err := clusterdiscovery.InitializeTestFramework(exutil.TestContext, config, testOpt.DryRun); err != nil {
+			return err
+		}
+		klog.V(4).Infof("Loaded test configuration: %#v", exutil.TestContext)
+
+		if err := upgrade.GatherPreUpgradeResourceCounts(); err != nil {
+			return errors.Wrap(err, "error gathering preupgrade resource counts")
+		}
+	}
+
+	// TODO this is called from run-upgrade and run-test.  At least one of these ought not need it.
+	return upgradeoptions.SetUpgradeGlobalsFromTestOptions(o.TestOptions)
+}
+
+func (o *RunUpgradeSuiteOptions) Run(ctx context.Context) error {
 	defer o.CloseFn()
 
 	// set globals so that helpers will create pods with the mapped images if we create them from this process.
@@ -53,7 +93,10 @@ func (o *RunSuiteOptions) Run(ctx context.Context) error {
 	// we cannot eliminate the env var usage until we convert run-test, which we may be able to do in a followup.
 	image.InitializeImages(o.FromRepository)
 
-	if err := verifyImages(); err != nil {
+	if err := imagesetup.VerifyTestImageRepoEnvVarUnset(); err != nil {
+		return err
+	}
+	if err := imagesetup.VerifyImages(); err != nil {
 		return err
 	}
 
@@ -67,16 +110,18 @@ func (o *RunSuiteOptions) Run(ctx context.Context) error {
 	// TODO fix the the upstream so that the AfterReadingAllFlags will properly check for either of the inputs having values.
 	k8simage.Init("")
 
+	if err := o.UpgradeTestPreSuite(); err != nil {
+		return err
+	}
+
 	o.GinkgoRunSuiteOptions.CommandEnv = o.TestCommandEnvironment()
 	if !o.GinkgoRunSuiteOptions.DryRun {
 		fmt.Fprintf(os.Stderr, "%s version: %s\n", filepath.Base(os.Args[0]), version.Get().String())
 	}
-	exitErr := o.GinkgoRunSuiteOptions.Run(o.Suite, "openshift-tests", false)
+	exitErr := o.GinkgoRunSuiteOptions.Run(o.Suite, "openshift-tests-upgrade", true)
 	if exitErr != nil {
 		fmt.Fprintf(os.Stderr, "Suite run returned error: %s\n", exitErr.Error())
 	}
 
-	// Special debugging carve-outs for teams is likely to age poorly.
-	clusterdiscovery.PrintStorageCapabilities(o.GinkgoRunSuiteOptions.Out)
 	return exitErr
 }
