+++ pkg/defaultinvariants/types.go
@@ -12,6 +12,7 @@ import (
 	"github.com/openshift/origin/pkg/invariants/disruptionserializer"
 	"github.com/openshift/origin/pkg/invariants/disruptionserviceloadbalancer"
 	"github.com/openshift/origin/pkg/invariants/intervalserializer"
+	"github.com/openshift/origin/pkg/invariants/knownimagechecker"
 	"github.com/openshift/origin/pkg/invariants/timelineserializer"
 	"github.com/openshift/origin/pkg/invariants/trackedresourcesserializer"
 	"github.com/openshift/origin/pkg/invariants/watchpods"
@@ -71,7 +72,6 @@ func newDisruptiveInvariants() invariants.InvariantRegistry {
 func newUniversalInvariants() invariants.InvariantRegistry {
 	invariantTests := invariants.NewInvariantRegistry()
 
-	// TODO add invariantTests here
 	invariantTests.AddInvariantOrDie("pod-lifecycle", "Test Framework", watchpods.NewPodWatcher())
 	invariantTests.AddInvariantOrDie("timeline-serializer", "Test Framework", timelineserializer.NewTimelineSerializer())
 	invariantTests.AddInvariantOrDie("interval-serializer", "Test Framework", intervalserializer.NewIntervalSerializer())
@@ -80,6 +80,7 @@ func newUniversalInvariants() invariants.InvariantRegistry {
 	invariantTests.AddInvariantOrDie("alert-summary-serializer", "Test Framework", alertserializer.NewAlertSummarySerializer())
 	invariantTests.AddInvariantOrDie("cluster-info-serializer", "Test Framework", clusterinfoserializer.NewClusterInfoSerializer())
 	invariantTests.AddInvariantOrDie("additional-events-collector", "Test Framework", additionaleventscollector.NewIntervalSerializer())
+	invariantTests.AddInvariantOrDie("known-image-checker", "Test Framework", knownimagechecker.NewEnsureValidImages())
 
 	return invariantTests
 }
