+++ pkg/test/ginkgo/cmd_runsuite.go
@@ -180,7 +180,6 @@ func (o *GinkgoRunSuiteOptions) Run(suite *TestSuite, junitSuiteName string, upg
 	}
 
 	syntheticEventTests := JUnitsForAllEvents{
-		o.SyntheticEventTests,
 		suite.SyntheticEventTests,
 	}
